Refactor to separate gaps and data requirements
