+++ src/helpers/ClauseResultsHelpers.ts
