+++ test/unit/queryFilters/interpretEquivalent.test.ts
@@ -1,5 +1,5 @@
 import { getELMFixture } from '../helpers/testHelpers';
-import * as QueryFilter from '../../../src/gaps/QueryFilterParser';
+import * as QueryFilter from '../../../src/helpers/elm/QueryFilterParser';
 import { ELMEquivalent } from '../../../src/types/ELMTypes';
 
 // to use as a library parameter for tests
