+++ test/unit/queryFilters/interpretIn.test.ts
@@ -1,6 +1,6 @@
 import { getELMFixture } from '../helpers/testHelpers';
 import { DateTime, Interval } from 'cql-execution';
-import * as QueryFilter from '../../../src/gaps/QueryFilterParser';
+import * as QueryFilter from '../../../src/helpers/elm/QueryFilterParser';
 import { ELMIn } from '../../../src/types/ELMTypes';
 import { DuringFilter } from '../../../src/types/QueryFilterTypes';
