+++ test/unit/queryFilters/interpretNot.test.ts
@@ -1,4 +1,4 @@
-import * as QueryFilter from '../../../src/gaps/QueryFilterParser';
+import * as QueryFilter from '../../../src/helpers/elm/QueryFilterParser';
 import { ELMNot } from '../../../src/types/ELMTypes';
 import { UnknownFilter } from '../../../src/types/QueryFilterTypes';
